Fixes 500 when creating an organization, add end-to-end tests
